Separated component checks from other semantic checks.
