+++ src/AstSemanticChecker.h
@@ -28,10 +28,6 @@ class AstAggregator;
 class AstArgument;
 class AstClause;
 class AstRelation;
-class AstComponent;
-class AstComponentScope;
-class AstComponentType;
-class AstComponentInit;
 class AstSrcLocation;
 class TypeBinding;
 class AstUnionType;
@@ -39,7 +35,6 @@ class AstRecordType;
 class AstType;
 class TypeEnvironment;
 class TypeAnalysis;
-class ComponentLookup;
 class PrecedenceGraph;
 class RecursiveClauses;
 
@@ -48,7 +43,7 @@ private:
     bool transform(AstTranslationUnit& translationUnit) override;
 
     static void checkProgram(ErrorReport& report, const AstProgram& program, const TypeEnvironment& typeEnv,
-            const TypeAnalysis& typeAnalysis, const ComponentLookup& componentLookup,
+            const TypeAnalysis& typeAnalysis,
             const PrecedenceGraph& precedenceGraph, const RecursiveClauses& recursiveClauses);
 
     static void checkAtom(ErrorReport& report, const AstProgram& program, const AstAtom& atom);
@@ -66,20 +61,6 @@ private:
             const AstRelation& relation, const RecursiveClauses& recursiveClauses);
     static void checkRules(ErrorReport& report, const TypeEnvironment& typeEnv, const AstProgram& program,
             const RecursiveClauses& recursiveClauses);
-
-    static const AstComponent* checkComponentNameReference(ErrorReport& report,
-            const AstComponent* enclosingComponent, const ComponentLookup& componentLookup,
-            const std::string& name, const AstSrcLocation& loc, const TypeBinding& binding);
-    static void checkComponentReference(ErrorReport& report, const AstComponent* enclosingComponent,
-            const ComponentLookup& componentLookup, const AstComponentType& type, const AstSrcLocation& loc,
-            const TypeBinding& binding);
-    static void checkComponentInit(ErrorReport& report, const AstComponent* enclosingComponent,
-            const ComponentLookup& componentLookup, const AstComponentInit& init, const TypeBinding& binding);
-    static void checkComponent(ErrorReport& report, const AstComponent* enclosingComponent,
-            const ComponentLookup& componentLookup, const AstComponent& component,
-            const TypeBinding& binding);
-    static void checkComponents(
-            ErrorReport& report, const AstProgram& program, const ComponentLookup& componentLookup);
 
     static void checkUnionType(ErrorReport& report, const AstProgram& program, const AstUnionType& type);
     static void checkRecordType(ErrorReport& report, const AstProgram& program, const AstRecordType& type);
